requirements.txt
Django
video_finder @git+git://github.com/MrKrautee/simple-youtube-video-finder.git
djangorestframework
markdown
django-filter
